feat: Add Event Route in JWT Token

+++ routes/events.py
@@ -30,13 +30,20 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 	return event
 @event_router.post("/new")
 async def create_event(body: Event, user: str=Depends(authenticate)) -> dict:
+	body.creator = user
 	await event_database.save(body)
 	return {
 		"message": "Event created successfully."
 	}
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str=Depends(authenticate)) -> Event:
+	event = await event_database.get(id)
+	if event.creator != user:
+		raise HTTPException(
+			status_code=status.HTTP_400_BAD_REQUEST,
+			detail="Operation not allowed"
+		)
 	updated_event = await event_database.update(id, body)
 	if not update_event:
 		raise HTTPException(
@@ -47,6 +54,12 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
 
 @event_router.delete("/{id}")
 async def delete_event(id: PydanticObjectId, user: str=Depends(authenticate)) -> dict:
+	event = await event_database.get(id)
+	if event.creator != user:
+		raise HTTPException(
+			status_code=status.HTTP_404_NOT_FOUND,
+			detail="Operation not allowed"
+		)
 	event = await event_database.delete(id)
 	if not event:
 		raise HTTPException(
